apps/processor/processor-generators/pyproject.toml
[project]
name = "processor-generators"
version = "0.1.0"
description = "Consolidated report generator service – PPTX, Excel generation and MCP AI agent interface."
requires-python = ">=3.12"
dependencies = [
    # Web framework & server
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    # gRPC
    "grpcio>=1.62",
    "grpcio-tools>=1.62",
    "grpcio-health-checking>=1.62",
    "protobuf>=4.25",
    # PPTX generation
    "python-pptx>=1.0.2",
    "Pillow>=10.0",
    # Excel generation
    "openpyxl>=3.1",
    # Chart rendering
    "matplotlib>=3.8",
    "numpy>=1.26",
    # MCP / AI
    "mcp>=1.0",
    "litellm>=1.30",
    "tiktoken>=0.6",
    "pyjwt[crypto]>=2.8",
    # Database
    "asyncpg>=0.29",
    # Azure Storage
    "azure-storage-blob>=12.19",
    # Observability
    "opentelemetry-api>=1.23",
    "opentelemetry-sdk>=1.23",
    "opentelemetry-instrumentation-grpc>=0.44b0",
    "opentelemetry-instrumentation-fastapi>=0.44b0",
    "opentelemetry-exporter-otlp>=1.23",
    # Dapr
    "dapr>=1.13",
    "dapr-ext-grpc>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "ruff>=0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 120
